Moved isAuthenticated to middleware
